Spring: Check current password in domain model
Changing the user's password requires the input of their current
password for verification. This didn't happen in the domain layer but
somewhere in the presentation layer.

Now that this is part of the domain there's no single space where
services needs to be injected with AOP magic (@Configurable) and
therefore the configuration becomes slightly simpler.